Fix vibe review issues
