+++ ydb/core/security/external_idp/test_utils/rsa_key_pair.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <util/generic/string.h>
+
+namespace NKikimr::NExternalIdpTestUtils {
+
+struct TRsaKeyPair {
+    TString PrivateKeyPem;
+    TString PublicKeyPem;
+    TString X5cBase64;
+};
+
+TRsaKeyPair GenerateRsaKeyPair();
+
+} // namespace NKikimr::NExternalIdpTestUtils
