+++ ydb/core/security/external_idp/test_utils/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    rsa_key_pair.cpp
+)
+
+PEERDIR(
+    library/cpp/string_utils/base64
+    contrib/libs/openssl
+)
+
+END()
